Finished setup_qm.py; Added submit.py
I finished up setup_qm.py to be workable. Options for processors,
memory and WFN still need to be added but the code does seem stable.
I also noticed that the code doesn't figure out which structure is
the minium structure when only writing a single com file for a given
maestro file. This would be a nice addition to the code. I tried
making sure the GaussCom class was general so that this can be used
when grabing structures from other sources other than Schrodinger.
I did have to use a mae_struct property in the class to remove any
dummy atoms, but this is very specific to how schrodinger treats
dummy atoms in their code.

Submit.py is a short script that I have used on a regular basis for
about a year. This script allows a Notre Dame CRC user to submit
crc jobs regardless of the type of job. This can also be used
for multiple files at one time and will submit them to the CRC
from the command line.