Added submit.py; finished setup_qm.py
submit.py
I have used this script for about a year and it lets a user submit
a job to the Notre Dame CRC. There are options to change processors,
queue type, and the command. This script will write the CRC *.sh and
submit it. Before using this change any defaults you want and test
it by commenting out the sp.call( line and uncommenting the print
statements that come after it.

setup_qm.py
This code should now be stable to be used. I still want to add the ability
to alter the number of processors, the amount of memory, and add in the
WFN option I mentioned last commit. It would also be nice to actually have
a for sure way to use the minimum energy structure when just writing one
Gaussian Command file from several structures in a schrodinger file. I have
attempted to make the GaussCom class as general as I could so that other
sources of structures can be used such as TINKER, etc. I did have to use
a mae_struct property for the GaussCom class to remove dummy atoms along
with matching substructures for frozen coordinates. I don't know of any
better way of doing this. The alternative would probably to have the user
know what the atom numbers they want frozen and to use that as command line
arguments for the frozen atoms. Additionally I use the structure to determine
elements for the basis set aswell. When we start using TINKER or other
file formats more then we will have to address this dependence on schrodinger.