Provide FF_row for data output
I added the 'ff_row' property to Datum so that we can reference the
ff_row easily for datum that reflect a single bonding interaction.
This will now print out the compare with an additional column with the
row value in the last column. I wonder if there would be a way to just
incorporate the Datum class from datatypes and the Bond (angle, torsion)
class in filetypes since they are almost redundant classes. I suppose
there is a benefit to a datum class having the sole classifcation as
the value of the datum.