New compare-Merge (#23)
* Update loop and compare

Instead of dealing with full arrays of data for reference and calculated
data, this will make a dictionary of data types for reference and calculated
data. The code also had two seperate functions for printing data and comparing
data, which is not consolidated into one function with the option of printing
data if necessary. I think there needs to be some additional support for
reference data and parameter tethering data.

* Modify modules to work with new compare.py

Compare.py was updated to return the total score and not the last
calculated score of an individual point.

gradient.py and simplex.py were updated to use the new function in
compare.py

* Fixed Label Bug.

Under some circumstances two data points that should be the same
interaction were being removed in trim_data(). This was because the
labels were different; the structure index was different. This is
probably due to compressing the number of MacroModel calculations for
a single structure resulting in a single output file with all of the
structures within that file. Using a regex I was able to match labels
only based on filename and atom invovled in the dihedral. This should
fix this problem now.

Gradient.py has some problems that I still need to address when handling
the new data dictionaries. For now I am just useing do_newton() since it
does not need to go through the data like the others do.

* Small fixes

* Torsion Label REGEX modified.

I had to use '\S' to capture all non-whitespace characters that may be
included in the filename. Additionally if a file has more than 9 structures
the regex wouldn't match so instead of '\d' I used '\d+'.

* Updated Logic for new_compare

Several things had to be modified to use the dictionaries that have
been set up in previous commits. The par_diff file is now written
in the same order as the compare files. The other modifications
just included reading data from the dictionaries rather than an
array of data.

* MacroModel MMO data gets sorted

This shouldn't really be to big of a problem, but I think the function
I added along with what is present in the Tinker class should be a
function within the structure class.

Structures with different filenames but with the same structure do not
print interactions to the *.mmo file fully sorted. It seems that they are
only sorted based on the index of one atom (bonds: first atom; angles: central atom; torsions: second atom).
I added additional logic that will sort the bonds, angles, and torsions of a structure
to a completely sorted list that is the same everytime.

* Bug when using gradient

Gradient was recalculating the score of the original force field
for all of the differentiation steps and thus reported a 1st and 2nd
derivative of 0.0. The appropriate logic had to be added to convert
the new calculated data into the datatype dictionary.

* Logging and Sort Torsions

Loggin was added to report all the data that was removed. Torsions,
I guess, are sorted in the *.mmo file based off of the lowest terminal
atom. I gave it similar logic to angles to sort based off of the internal
atoms first.

* Sorting Torsions

Read previous commit.

* Bug Fixes for previous commit

* Additional Bugs in Gradient

The parameter differentiation file was not being written correctly with
respect to the differentiation force fields. Just added logic to write
the trimed and sorted data correctly.

* Label to Attr for param tethering in reference file.

* Added support to write com file to different directory.

User can now specify directory that the gaussian *.com file should
be written in.