Added more substrates
Added more aryl substrates bound to a metal, with the conformational search parameters, to substrates/aryl_alkenyl/.
Updated the chemdraw for aryl_alkenyl-substrates.cdx to include the new aryl substrates.
Created a new directory in substrates for enones. This contains 30 enone substrates with conformational search parameters as well as a chemdraw to describe the structures.