Added nonbonded cutoffs
Added the EXNB cutoff used to set the cutoffs for nonbonded interactions.
This was needed to ensure all intractions are being gathered even at long
distances. This shouldn't be too much of a problem for virtual screening
purposes, but in the case of much larger systems this may have a drain
on computational speed.