Fixes three small bugs
Fix for printing ParamError exception messages

BaseException.message was deprecated in python 2.5 and has been removed
starting with python3. Exception has __str__ and __repr__ methods implemented
such that you can print the Exception message by calling str() on the
caught exception.

Adds a param check after optimization completes

Adds a small check at the end of the optimization loop to see if any
parameters are at either of its limits and prints a warning to the
user if they are. Also removes resetting the step size for a param
to 0.1 if the step is 0.0. Fixes #51

Fixes a small bug that Jess found

I made a copy/paste bug here that Jess found and helped me to
resolve. More reason we need to get Q2MM packaged properly and
add some testing to run before pull requests can be merged so that
I don't push bugs to the master repo and ruin everyone's workflow.