Fix bug min value in Hessian matrix being replaced
When the option to use the Jaguar Hessian was specified in
`calculate.py`, the minimum value of the Hessian was being replaced.
This is a bug. In this commit, I remove the line of code that replaces
the minimum value of the Hessian.

To invert the Hessian, the Hessian must first be converted into its
eigenmodes. Then most negative eigenmode should then be replaced. Then
the Hessian should be reformed. This process takes place when `--invert`
is specified IN ADDITION TO the bug mentioned above.

For a full discusion on this issue, please refer to issue #57. Closes
issue #57.